requirements.txt
flask==2.2.5
werkzeug==2.2.3
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
pillow==9.5.0
opencv-python==4.7.0.72
pennylane==0.31.0
pennylane-qiskit==0.28.0
qiskit==0.44.0
seaborn==0.12.2
jupyterlab==4.0.3
ipykernel==6.25.0
